api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pyyaml>=6.0
python-frontmatter>=1.1.0
tiktoken>=0.7.0
httpx>=0.27.0
